# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.MapResponse:
    properties:
      map_url:
        description: Публичный URL карты мероприятия
        example: http://localhost:8080/static/maps/a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11_map.jpg
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Сервис для управления событиями, картами и точками интереса.
  title: Event Designer API
  version: "1.0"
paths:
...
  /admin/events/{eventID}/map:
    delete:
      description: Удаляет карту мероприятия и соответствующий файл из хранилища.
      parameters:
      - description: ID мероприятия (UUID)
        in: path
        name: eventID
        required: true
        type: string
      responses:
        "204":
          description: Карта успешно удалена (No Content)
        "400":
          description: Invalid event ID format in URL
          schema:
            type: string
        "404":
          description: Event map not found.
          schema:
            type: string
        "500":
          description: Failed to delete event map
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Удалить карту мероприятия
      tags:
      - Maps
...
  /admin/events/map:
    put:
      consumes:
      - multipart/form-data
      description: Загружает новый файл карты или заменяет существующий для мероприятия.
      parameters:
      - description: ID мероприятия (UUID)
        in: formData
        name: event_id
        required: true
        type: string
      - description: Файл карты (изображение)
        in: formData
        name: map_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Карта успешно создана (StatusCreated)
        "204":
          description: Карта успешно обновлена (No Content)
        "400":
          description: Invalid event ID format / Missing form data
          schema:
            type: string
        "500":
          description: Failed to upload file / Failed to create/update event map in
            DB
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Загрузить или обновить карту мероприятия
      tags:
      - Maps
...
  /events/{eventID}/map:
    get:
      description: Возвращает URL общедоступной карты мероприятия.
      parameters:
      - description: ID мероприятия (UUID)
        in: path
        name: eventID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: URL карты
          schema:
            $ref: '#/definitions/models.MapResponse'
        "400":
          description: Invalid event ID format in URL
          schema:
            type: string
        "404":
          description: Event map not found.
          schema:
            type: string
        "500":
          description: Failed to fetch event map
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Получить карту мероприятия
      tags:
      - Maps
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"